divide components to sub components
