+++ src/routes/SearchPage/index.js
@@ -1,28 +1,15 @@
-import { useCallback, useContext, useEffect, useRef, useState } from 'react'
+import { useCallback, useContext, useEffect, useState } from 'react'
 import { Outlet, useNavigate } from 'react-router-dom'
 
-import { Autocomplete, useJsApiLoader } from '@react-google-maps/api'
+import { useJsApiLoader } from '@react-google-maps/api'
 
-import {
-  Box,
-  FormControl,
-  FormLabel,
-  IconButton,
-  InputLabel,
-  Stack,
-  Typography,
-} from '@mui/material'
-import { Radio, RadioGroup, Sheet, radioClasses } from '@mui/joy'
-import { Clear } from '@mui/icons-material'
-import { grey } from '@mui/material/colors'
-import { visuallyHidden } from '@mui/utils'
+import { Box, FormControl, Stack, Typography } from '@mui/material'
 
 import Map from '../../components/Map'
 import { DirectionsContext } from '../../contexts/directions.context'
-import { MelbourneBounds } from '../../constants/map.constant'
-import { ModeList, IconList } from '../../constants/iconlist.constant'
 
-import './index.css'
+import SearchBox from '../../components/SearchBox'
+import SelectTravelMode from '../../components/SelectTravelMode'
 
 const GOOGLE_API_KEY = process.env.REACT_APP_GOOGLE_MAPS_API_KEY
 
@@ -30,11 +17,6 @@ const SearchPage = () => {
   const { directions, setDirections } = useContext(DirectionsContext)
   // To remove a warning
   const [libraries] = useState(['places'])
-  const [origAutoComplete, setOrigAutoComplete] = useState(null)
-  const [destAutoComplete, setDestAutoComplete] = useState(null)
-
-  let originRef = useRef()
-  let destinationRef = useRef()
 
   let navigate = useNavigate()
 
@@ -74,52 +56,6 @@ const SearchPage = () => {
     }
   }, [calculateRoute, isLoaded, directions])
 
-  const handleTravelMode = (event) => {
-    const travelMode = event.target.value
-    setDirections({ ...directions, travelMode: travelMode, results: null })
-  }
-
-  const handlePlaceChanged = (autoComplete, mode) => {
-    const place = autoComplete.getPlace()
-    if (!place.place_id) return
-
-    if (mode === 'ORIG') {
-      setDirections({
-        ...directions,
-        origPlaceId: place.place_id,
-        origin: originRef.current.value,
-        results: null,
-      })
-    } else {
-      setDirections({
-        ...directions,
-        destPlaceId: place.place_id,
-        destination: destinationRef.current.value,
-        results: null,
-      })
-    }
-  }
-
-  const clearAddress = (mode) => {
-    if (mode === 'ORIG') {
-      originRef.current.value = ''
-      setDirections({
-        ...directions,
-        origin: '',
-        origPlaceId: null,
-        results: null,
-      })
-    } else {
-      destinationRef.current.value = ''
-      setDirections({
-        ...directions,
-        destination: '',
-        destPlaceId: null,
-        results: null,
-      })
-    }
-  }
-
   if (!isLoaded)
     return (
       <Typography
@@ -157,138 +93,9 @@ const SearchPage = () => {
           }}
         >
           <FormControl>
-            <Stack position="relative">
-              <InputLabel htmlFor="start" sx={visuallyHidden}>
-                Start address
-              </InputLabel>
-              <Autocomplete
-                bounds={MelbourneBounds}
-                fields={['place_id']}
-                onLoad={(autoComplete) => {
-                  setOrigAutoComplete(autoComplete)
-                }}
-                onPlaceChanged={() =>
-                  handlePlaceChanged(origAutoComplete, 'ORIG')
-                }
-                options={{ strictBounds: true }}
-              >
-                <input
-                  className="search-input"
-                  type="text"
-                  name="start"
-                  id="start"
-                  placeholder="Start"
-                  aria-label="start-address"
-                  required
-                  defaultValue={directions.origin}
-                  ref={originRef}
-                />
-              </Autocomplete>
-              <IconButton
-                disableRipple
-                sx={{
-                  position: 'absolute',
-                  right: '2px',
-                  top: '4px',
-                  bgcolor: 'white',
-                  ':hover': { color: 'black' },
-                }}
-                onClick={() => clearAddress('ORIG')}
-              >
-                <Clear />
-              </IconButton>
-            </Stack>
-            <Stack position="relative">
-              <InputLabel htmlFor="end" sx={visuallyHidden}>
-                End address
-              </InputLabel>
-              <Autocomplete
-                bounds={MelbourneBounds}
-                fields={['place_id']}
-                onLoad={(autoComplete) => {
-                  setDestAutoComplete(autoComplete)
-                }}
-                onPlaceChanged={() =>
-                  handlePlaceChanged(destAutoComplete, 'DEST')
-                }
-                options={{ strictBounds: true }}
-              >
-                <input
-                  className="search-input"
-                  type="text"
-                  name="end"
-                  id="end"
-                  placeholder="End"
-                  aria-label="end-address"
-                  defaultValue={directions.destination}
-                  ref={destinationRef}
-                />
-              </Autocomplete>
-              <IconButton
-                disableRipple
-                sx={{
-                  position: 'absolute',
-                  right: '2px',
-                  top: '4px',
-                  bgcolor: 'white',
-                  ':hover': { color: 'black' },
-                }}
-                onClick={() => clearAddress('DEST')}
-              >
-                <Clear />
-              </IconButton>
-            </Stack>
-            <FormLabel id="travelmode-radio-buttons-label" sx={visuallyHidden}>
-              Select Travel Mode
-            </FormLabel>
-            <RadioGroup
-              row
-              sx={{ gap: 2, mt: 1 }}
-              aria-labelledby="travelmode-radio-buttons-label"
-              name="travelmode-radio-buttons-group"
-              defaultValue="DRIVING"
-              value={directions.travelMode}
-              onChange={handleTravelMode}
-            >
-              {ModeList.map((mode) => (
-                <Sheet
-                  key={mode}
-                  sx={{
-                    position: 'relative',
-                    width: 40,
-                    height: 40,
-                    flexShrink: 1,
-                    borderRadius: '50%',
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    [`& .${radioClasses.checked}`]: {
-                      [`& .${radioClasses.label}`]: {
-                        color: 'rgb(46, 145, 39)',
-                      },
-                      [`& .${radioClasses.action}`]: {
-                        '--variant-borderWidth': '2px',
-                        borderColor: 'rgb(46, 145, 39)',
-                      },
-                    },
-                    [`& .${radioClasses.action}.${radioClasses.focusVisible}`]:
-                      {
-                        outlineWidth: '2px',
-                      },
-                  }}
-                >
-                  <Radio
-                    sx={{ color: grey[500] }}
-                    overlay
-                    disableIcon
-                    value={mode}
-                    label={IconList[mode]}
-                    name={mode}
-                    componentsProps={{ input: { 'aria-label': { mode } } }}
-                  />
-                </Sheet>
-              ))}
-            </RadioGroup>
+            <SearchBox mode="start" />
+            <SearchBox mode="end" />
+            <SelectTravelMode />
           </FormControl>
         </Stack>
         <Outlet />
